requirements.txt
django
django-rest-framework
requests
drf_yasg
gunicorn